+++ LaserScan.Core.NetStandart/Models/MaterialModel.cs
@@ -28,7 +28,7 @@ namespace LaserScan.Core.NetStandart.Models
     public class CameraDelta:BindableBase
     {
         public string CameraId { get; set; }
-        public sbyte[] Deltas { get; set; }
+        public double[] Deltas { get; set; }
 
         private byte _upThreshhold;
         public byte UpThreshhold
