+++ Laserscan.IServices/Math/ICalibrateService.cs
@@ -4,7 +4,7 @@
     {
         (double[], int[]) Calibrate(byte[] data);
         (int, double) CalibrateExposureTimeRaw(byte[] data, int currentExposure);
-        sbyte[] CalibrateRaw(byte[] data);
+        double[] CalibrateRaw(byte[] data);
         sbyte[] DefaultCalibration(double[] p, int count);
     }
 }
